requirements.txt
flask==2.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas
openai